pyproject.toml
[project]
name = "evommagent"
version = "0.1.0"
description = "A flexible framework for building multi-modal agents with tool support"
authors = [
    {name = "Wang Shijian"}
]
readme = "README.md"
requires-python = ">=3.11"
keywords = ["agent", "multimodal", "tool-use", "reasoning", "llm"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "jinja2>=3.0.0",
    "jsonschema>=4.0.0",
    "pillow>=10.0.0",
    "httpx>=0.24.0",
    "openai>=1.0.0",
    "requests>=2.31.0",
    "tqdm>=4.65.0",
    "python-dotenv>=1.0.0",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
]
vl = [
    "qwen-vl-utils==0.0.14",
    "decord>=0.6.0",
    "torchcodec>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agent", "tool", "api"]

